inventory_optimization_bundle/agent/requirements.txt
# Agent Framework dependencies
databricks-agents>=0.16.0
databricks-sdk>=0.15.0
mlflow>=2.20.2
pydantic>=2.0.0
